POKEMON: Refactor project logic
